Beta
RepuBikla is a tool in development for the creation and centralization
of data on non-motorized mobility in Mexico and elsewhere in the world.

It aims to complement insufficient official information on non motorized
mobility and give a citizen approach with a database being built and
strengthened progressively. The result is an assessment of the volume
and diversity of cyclists commuting, intermodal, mobility group,
infrastructure, cyclist accidents and insecurity in public space.

Decision makers and civil society are equipped with a tool to promote
and design better conditions for sustainable and safe mobility.

RepuBikla is stage funding. It is complemented by a smartphone
application that will help tools to the rider. The data collected will
create and enhance optimal routing algorithms and other functions used
by other applications using APIs. Also will complement counts.